+++ examples/rl_gsm8k/utils.py
@@ -12,16 +12,11 @@ import numpy as np
 import psutil
 import requests
 import torch
-from datasets import load_dataset
-from omegaconf import DictConfig
 from tenacity import retry, stop_after_attempt, wait_exponential
-from tqdm import tqdm
 from transformers import PreTrainedTokenizer
 
 from tapeagents.config import is_debug_mode
 from tapeagents.llms import LLMOutput, Prompt
-
-from .deepseek_math_eval.process_utils import process_eurus_test, process_gsm8k_test, process_math_test
 
 logger = logging.getLogger(__name__)
 
@@ -427,34 +422,3 @@ def get_tokens_from_hf_tokenizer(tokenizer: PreTrainedTokenizer | None, prompt: 
     output_token_ids = text_token_ids[len(prompt_token_ids) :]
     output_tokens = [tokenizer.decode(output_token_id) for output_token_id in output_token_ids]
     return output_tokens
-
-
-def load_datasets(cfg: DictConfig) -> Tuple[list, list]:
-    match cfg.dataset_name:
-        case "math":
-            train_dataset_long_name = test_dataset_long_name = "hendrycks/competition_math"
-            process_fn = process_math_test
-            builder_config = "main"
-        case "gsm8k":
-            train_dataset_long_name = test_dataset_long_name = "openai/gsm8k"
-            process_fn = process_gsm8k_test
-            builder_config = "main"
-        case "eurus":
-            train_dataset_long_name = "PRIME-RL/Eurus-2-RL-Data"
-            test_dataset_long_name = "alexpiche/math_test_cleaned"
-            process_fn = process_eurus_test
-            builder_config = "default"
-        case _:
-            raise ValueError(f"Unknown dataset: {cfg.dataset_name}")
-
-    train_dataset = load_dataset(train_dataset_long_name, builder_config, split="train", trust_remote_code=True)
-    test_dataset = load_dataset(test_dataset_long_name, builder_config, split="test", trust_remote_code=True)
-    train_samples = [
-        process_fn(s) for s in tqdm(train_dataset, desc="Processing train samples") if process_fn(s) is not None
-    ]
-    test_samples = [
-        process_fn(s) for s in tqdm(test_dataset, desc="Processing test samples") if process_fn(s) is not None
-    ]
-    logger.info(f"Loaded {len(train_samples)} training samples")
-    logger.info(f"Loaded {len(test_samples)} test samples")
-    return train_samples, test_samples
